Move Handlers to their own folders
