+++ handlers/AuthHandler/index.js
@@ -3,16 +3,16 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 export class AuthHandler extends React.Component {
-  static get propTypes() {
-    return {
-      dispatch: PropTypes.func,
-    };
-  }
-
   componentDidMount() {
     this.props.dispatch({
       type: 'getAuth',
     });
+  }
+
+  static get propTypes() {
+    return {
+      dispatch: PropTypes.func,
+    };
   }
 
   render() {
